+++ frontend/public/changelog.md
@@ -1,5 +1,9 @@
 # Riwayat Perubahan
 
+## 2026-03-04 Rabu
+
+- Membuat halaman redaksi untuk melihat statistik [Susun Kata Bebas](/gim/susun-kata/bebas)
+
 ## 2026-03-03 Selasa
 
 - Menampilkan kategori tagar di [Kamus](/kamus), misalnya *[peng-](/kamus/tagar/peng)*